add placeholders window
