+++ stfu-gui/src/main/resources/bundles/scenes/templates/add-placeholder-videos_de.properties
@@ -1,0 +1,9 @@
+main.description=Gib hier bitte den Dateinamen des Videos, dessen Veröffentlichungsdetails du planen willst, ein.&#13;Dabei kannst du die Dateiendung (z. B. .mkv oder .mp4) weglassen, es funktioniert aber auch mit der Endung.
+main.filename=Dateiname:
+multi.title=Mehrere Dateien hinzufügen (optional)
+multi.description=Du kannst mehrere Videos, deren Dateiname sich nur durch eine Zahl unterscheidet, hinzufügen.&#13;Trage einen Dateinamen oben ein, ersetze die Zahl durch {0} und fülle die drei Felder unten aus.&#13;Anschließend wird {0} durch alle Zahlen ersetzt und die Einträge werden hinzugefügt.&#13;&#13;Wenn eine Zahl weniger Ziffern hat, als bei Stellen angegeben wurde,&#13;wird mit führenden Nullen aufgefüllt.&#13;42 wird bei 3 eingestellten Stellen zu 042.
+multi.lowest=Niedrigste Zahl:
+multi.highest=Höchste Zahl:
+multi.leadingzeroes=Mindestanzahl Stellen:
+buttons.save=Speichern
+buttons.cancel=Abbrechen
