+++ stfu-gui/src/main/resources/bundles/scenes/templates/add-placeholder-videos_en.properties
@@ -1,0 +1,9 @@
+main.description=Please enter the filename of a video being planned for upload.&#13;You don't need to provide the extension (e. g. .mkv oder .mp4), but you can if you want.
+main.filename=Filename:
+multi.title=Add several files (optional)
+multi.description=If you have several filenames which only differ by a number, you can add them at once.&#13;To add several filenames at once, enter one of them, replace the number by {0} and enter the three fields below.&#13;{0} will be replaced by all numbers inside the given limit.&#13;&#13;Leading Zeroes will be added if a number has less digits than the amount set on "Minimum amount of digits".&#13;The number 42 will be added as 042 with minimum amount set to 3.
+multi.lowest=Smallest number:
+multi.highest=Greatest number:
+multi.leadingzeroes=Minimum amount of digits:
+buttons.save=Save
+buttons.cancel=Cancel
