+++ stfu-gui/src/main/resources/bundles/scenes/templates/templates_de.properties
@@ -69,3 +69,4 @@ template.edit.tab.twitter=Twitter
 template.edit.tab.twitter.header=Veröffentlichung auf Twitter
 template.edit.tab.scripting=Java-Scripting (Expertenmodus)
 template.edit.tab.scripting.header=Java-Scripting (Expertenmodus)
+subwindows.addplaceholdervideos=Dateien für Platzhalter hinzufügen
