+++ stfu-gui/src/main/resources/bundles/scenes/templates/templates_en.properties
@@ -69,3 +69,4 @@ template.edit.tab.twitter=Twitter
 template.edit.tab.twitter.header=Publication on Twitter
 template.edit.tab.scripting=Java-Scripting (Expert mode)
 template.edit.tab.scripting.header=Java-Scripting (Expert mode)
+subwindows.addplaceholdervideos=Add files for placeholders
